+++ src/routes/(station)/+page.server.ts
@@ -1,0 +1,7 @@
+import type { PageServerLoad } from './$types';
+
+export const load = (({}) => {}) satisfies PageServerLoad;
+
+export const prerender = true;
+export const ssr = true;
+export const csr = true;