+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2023-06-17
+
+#### Changed
+
+- site theme and name are inline now (Aerlon)
+- adjused home page (station) to use its own layout
+  - very much a WIP, goal is to make it look like the site is in the air at an airship station
+
+#### Added
+
+- top development versioning bar
+  - this will be removed when the core website is around version 1.0
+
 ## [0.4.8] - 2023-06-14
 
 #### Changed